+++ requirements-test.txt
@@ -1,3 +1,4 @@
 pycodestyle
 pytest
 pytest-cov
+pytest-mock
