+++ smbprotocol/__init__.py
@@ -18,34 +18,6 @@ logger.addHandler(NullHandler())
 MAX_PAYLOAD_SIZE = 65536
 
 
-class Commands(object):
-    """
-    [MS-SMB2] v53.0 2017-09-15
-
-    2.2.1.2 SMB2 Packet Header - SYNC Command
-    The command code of an SMB2 packet, it is used in the packet header.
-    """
-    SMB2_NEGOTIATE = 0x0000
-    SMB2_SESSION_SETUP = 0x0001
-    SMB2_LOGOFF = 0x0002
-    SMB2_TREE_CONNECT = 0x0003
-    SMB2_TREE_DISCONNECT = 0x0004
-    SMB2_CREATE = 0x0005
-    SMB2_CLOSE = 0x0006
-    SMB2_FLUSH = 0x0007
-    SMB2_READ = 0x0008
-    SMB2_WRITE = 0x0009
-    SMB2_LOCK = 0x000A
-    SMB2_IOCTL = 0x000B
-    SMB2_CANCEL = 0x000C
-    SMB2_ECHO = 0x000D
-    SMB2_QUERY_DIRECTORY = 0x000E
-    SMB2_CHANGE_NOTIFY = 0x000F
-    SMB2_QUERY_INFO = 0x0010
-    SMB2_SET_INFO = 0x0011
-    SMB2_OPLOCK_BREAK = 0x0012
-
-
 class Dialects(object):
     """
     [MS-SMB2] v53.0 2017-09-15
