+++ tests/conftest.py
@@ -14,6 +14,145 @@ from smbclient.shutil import (
     rmtree,
 )
 
+from smbprotocol.dfs import (
+    DFSReferralResponse,
+)
+
+DOMAIN_NAME = 'domain.test'
+DOMAIN_REFERRAL = DFSReferralResponse()
+DOMAIN_REFERRAL.unpack(b"\x00\x00"
+                       b"\x02\x00"
+                       b"\x00\x00\x00\x00"
+                       b"\x03\x00"
+                       b"\x12\x00"
+                       b"\x00\x00"
+                       b"\x00\x00"
+                       b"\x58\x02\x00\x00"
+                       b"\x24\x00"
+                       b"\x00\x00"
+                       b"\x00\x00"
+                       b"\x03\x00"
+                       b"\x12\x00"
+                       b"\x00\x00"
+                       b"\x02\x00"
+                       b"\x58\x02\x00\x00"
+                       b"\x22\x00"
+                       b"\x00\x00"
+                       b"\x00\x00"
+                       b"\x5C\x00\x44\x00\x4F\x00\x4D\x00"
+                       b"\x41\x00\x49\x00\x4E\x00\x00\x00"
+                       b"\x5C\x00\x64\x00\x6F\x00\x6D\x00"
+                       b"\x61\x00\x69\x00\x6E\x00\x2E\x00"
+                       b"\x74\x00\x65\x00\x73\x00\x74\x00"
+                       b"\x00\x00")
+
+DC_REFERRAL = DFSReferralResponse()
+DC_REFERRAL.unpack(b"\x00\x00"
+                   b"\x01\x00"
+                   b"\x00\x00\x00\x00"
+                   b"\x03\x00"
+                   b"\x12\x00"
+                   b"\x00\x00"
+                   b"\x02\x00"
+                   b"\x58\x02\x00\x00"
+                   b"\x12\x00"
+                   b"\x02\x00"
+                   b"\x2C\x00"
+                   b"\x5C\x00\x64\x00\x6F\x00\x6D\x00"
+                   b"\x61\x00\x69\x00\x6E\x00\x2E\x00"
+                   b"\x74\x00\x65\x00\x73\x00\x74\x00"
+                   b"\x00\x00"
+                   b"\x5C\x00\x44\x00\x43\x00\x30\x00"
+                   b"\x31\x00\x2E\x00\x64\x00\x6F\x00"
+                   b"\x6D\x00\x61\x00\x69\x00\x6E\x00"
+                   b"\x2E\x00\x74\x00\x65\x00\x73\x00"
+                   b"\x74\x00\x00\x00"
+                   b"\x5C\x00\x44\x00\x43\x00\x30\x00"
+                   b"\x31\x00\x2E\x00\x64\x00\x6F\x00"
+                   b"\x6D\x00\x61\x00\x69\x00\x6E\x00"
+                   b"\x2E\x00\x74\x00\x65\x00\x73\x00"
+                   b"\x74\x00\x32\x00\x00\x00")
+
+
+ROOT_REFERRAL = DFSReferralResponse()
+ROOT_REFERRAL.unpack(b"\x20\x00"
+                     b"\x01\x00"
+                     b"\x03\x00\x00\x00"
+                     b"\x04\x00"
+                     b"\x22\x00"
+                     b"\x01\x00"
+                     b"\x04\x00"
+                     b"\x2C\x01\x00\x00"
+                     b"\x22\x00"
+                     b"\x44\x00"
+                     b"\x66\x00"
+                     b"\x00\x00\x00\x00\x00\x00\x00\x00"
+                     b"\x00\x00\x00\x00\x00\x00\x00\x00"
+                     b"\x5C\x00\x64\x00\x6F\x00\x6D\x00"
+                     b"\x61\x00\x69\x00\x6E\x00\x2E\x00"
+                     b"\x74\x00\x65\x00\x73\x00\x74\x00"
+                     b"\x5C\x00\x64\x00\x66\x00\x73\x00"
+                     b"\x00\x00\x5C\x00\x64\x00\x6F\x00"
+                     b"\x6D\x00\x61\x00\x69\x00\x6E\x00"
+                     b"\x2E\x00\x74\x00\x65\x00\x73\x00"
+                     b"\x74\x00\x5C\x00\x64\x00\x66\x00"
+                     b"\x73\x00\x00\x00\x5C\x00\x53\x00"
+                     b"\x45\x00\x52\x00\x56\x00\x45\x00"
+                     b"\x52\x00\x32\x00\x30\x00\x31\x00"
+                     b"\x32\x00\x52\x00\x32\x00\x2E\x00"
+                     b"\x44\x00\x4F\x00\x4D\x00\x41\x00"
+                     b"\x49\x00\x4E\x00\x2E\x00\x54\x00"
+                     b"\x45\x00\x53\x00\x54\x00\x5C\x00"
+                     b"\x64\x00\x66\x00\x73\x00\x00\x00")
+
+
+TARGET_REFERRAL = DFSReferralResponse()
+TARGET_REFERRAL.unpack(b"\x26\x00"
+                       b"\x02\x00"
+                       b"\x02\x00\x00\x00"
+                       b"\x04\x00"
+                       b"\x22\x00"
+                       b"\x00\x00"
+                       b"\x04\x00"
+                       b"\x08\x07\x00\x00"
+                       b"\x44\x00"
+                       b"\x6C\x00"
+                       b"\x94\x00"
+                       b"\x00\x00\x00\x00\x00\x00\x00\x00"
+                       b"\x00\x00\x00\x00\x00\x00\x00\x00"
+                       b"\x04\x00"
+                       b"\x22\x00"
+                       b"\x00\x00"
+                       b"\x00\x00"
+                       b"\x08\x07\x00\x00"
+                       b"\x22\x00"
+                       b"\x4A\x00"
+                       b"\x9C\x00"
+                       b"\x00\x00\x00\x00\x00\x00\x00\x00"
+                       b"\x00\x00\x00\x00\x00\x00\x00\x00"
+                       b"\x5C\x00\x64\x00\x6F\x00\x6D\x00"
+                       b"\x61\x00\x69\x00\x6E\x00\x2E\x00"
+                       b"\x74\x00\x65\x00\x73\x00\x74\x00"
+                       b"\x5C\x00\x64\x00\x66\x00\x73\x00"
+                       b"\x5C\x00\x64\x00\x63\x00\x00\x00"
+                       b"\x5C\x00\x64\x00\x6F\x00\x6D\x00"
+                       b"\x61\x00\x69\x00\x6E\x00\x2E\x00"
+                       b"\x74\x00\x65\x00\x73\x00\x74\x00"
+                       b"\x5C\x00\x64\x00\x66\x00\x73\x00"
+                       b"\x5C\x00\x64\x00\x63\x00\x00\x00"
+                       b"\x5C\x00\x64\x00\x63\x00\x30\x00"
+                       b"\x31\x00\x2E\x00\x64\x00\x6F\x00"
+                       b"\x6D\x00\x61\x00\x69\x00\x6E\x00"
+                       b"\x2E\x00\x74\x00\x65\x00\x73\x00"
+                       b"\x74\x00\x5C\x00\x63\x00\x24\x00"
+                       b"\x00\x00\x5C\x00\x73\x00\x65\x00"
+                       b"\x72\x00\x76\x00\x65\x00\x72\x00"
+                       b"\x32\x00\x30\x00\x31\x00\x39\x00"
+                       b"\x2E\x00\x64\x00\x6F\x00\x6D\x00"
+                       b"\x61\x00\x69\x00\x6E\x00\x2E\x00"
+                       b"\x74\x00\x65\x00\x73\x00\x74\x00"
+                       b"\x5C\x00\x63\x00\x24\x00\x00\x00")
+
 
 @pytest.fixture(scope='module')
 def smb_real():
