+++ tests/test_structure.py
@@ -16,12 +16,15 @@ from datetime import (
 )
 
 from smbprotocol import (
-    Commands,
     Dialects,
 )
 
 from smbprotocol.connection import (
     Capabilities,
+)
+
+from smbprotocol.header import (
+    Commands,
 )
 
 from smbprotocol.structure import (
@@ -1262,7 +1265,7 @@ class TestEnumField(object):
         with pytest.raises(ValueError) as exc:
             field.set_value(0x13)
         assert str(exc.value) == "Enum value 19 does not exist in enum type " \
-                                 "<class 'smbprotocol.Commands'>"
+                                 "<class 'smbprotocol.header.Commands'>"
 
 
 class TestFlagField(object):
